+++ src/FastTravel/Framework/ModConstants.cs
@@ -1,0 +1,8 @@
+namespace Pathoschild.TheLongDarkMods.FastTravel.Framework;
+
+/// <summary>The mod constants.</summary>
+internal static class ModConstants
+{
+    /// <summary>The maximum number of 'favorite' destinations which can be assigned to a keybind.</summary>
+    public const int MaxFavorites = 9;
+}